cli/requirements.txt
typer[all]>=0.9.0
supabase>=2.3.0
httpx>=0.26.0
rich>=13.0.0
python-dotenv>=1.0.0
